pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Personal Assistant agent built using the Strands Agents framework. Agent as a tool pattern is replicated here."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.28.0",
    "strands-agents-tools>=0.2.21",
    "urllib3>=2.6.0",
]
